ec2: Consolidate 'filter.go' into 'filters.go'.
